# backend/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: CashX API
  version: 1.0.0
  description: Partner platform API. All money in kopecks (int64), rates in basis points.
servers:
  - url: /api/v1
paths:
  /auth/register:
    post:
      operationId: authRegister
      summary: Register a partner account (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/UserMe'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
...
  /auth/password-reset/request:
    post:
      operationId: authPasswordResetRequest
      summary: Request password reset (always 202; token returned in dev)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  reset_token:
                    type: string
                    nullable: true
  /auth/password-reset/confirm:
    post:
      operationId: authPasswordResetConfirm
      summary: Confirm password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, new_password]
              properties:
                token:
                  type: string
                new_password:
                  type: string
      responses:
        '204':
          description: No content
        '400':
          $ref: '#/components/responses/Error'
...
components:
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        referral_code:
          type: string
    UserMe:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [partner, staff]
        is_approved:
          type: boolean
        is_blocked:
          type: boolean
        partner:
          type: object
          nullable: true
          properties:
            referral_code:
              type: string
            telegram_user_id:
              type: integer
              format: int64
              nullable: true
        staff:
          type: object
          nullable: true
          properties:
            roles:
              type: array
              items:
                type: string
...